allUnits_early_window_matrix()
this finds the firing rate in a given window around given trigger times. you can provide a list of trigger times, most likely representing each trial for a unit. you can then shift the start time forward a given amount of time, and then select a time window to forward to find the firing rate of. NOTE: that you need to calculate the number of fr to use to find the proper avg. for example here im using 40ms window with 5ms bins, there for 40/5 =8, to calculate the fr in your window you need to only use the first 8 entries in the fr list of lists